+++ code/core/src/main/java/org/betonquest/betonquest/kernel/processor/QuestProcessor.java
@@ -61,7 +61,20 @@ public abstract class QuestProcessor<I extends Identifier, T> implements FastSta
                           final IdentifierFactory<I> identifierFactory, final String readable, @SourceId final String internal) {
         this.log = log;
         this.identifierFactory = identifierFactory;
-        this.values = new HashMap<>();
+        this.values = new HashMap<>() {
+            @Override
+            public T put(final I key, final T value) {
+                if (key == null) {
+                    log.error("Key of %s with value '%s' may not be null".formatted(readable, value));
+                    return null;
+                }
+                if (value == null) {
+                    log.error("Value of %s with key '%s' may not be null".formatted(readable, key));
+                    return null;
+                }
+                return super.put(key, value);
+            }
+        };
         this.readable = readable;
         this.internal = internal;
     }