+++ code/core/src/main/java/org/betonquest/betonquest/quest/condition/point/GlobalPointConditionFactory.java
@@ -1,8 +1,6 @@
 package org.betonquest.betonquest.quest.condition.point;
 
 import org.betonquest.betonquest.api.QuestException;
-import org.betonquest.betonquest.api.instruction.Argument;
-import org.betonquest.betonquest.api.instruction.FlagArgument;
 import org.betonquest.betonquest.api.instruction.Instruction;
 import org.betonquest.betonquest.api.quest.condition.NullableConditionAdapter;
 import org.betonquest.betonquest.api.quest.condition.PlayerCondition;
@@ -14,12 +12,7 @@ import org.betonquest.betonquest.database.GlobalData;
 /**
  * Factory to create global point conditions from {@link Instruction}s.
  */
-public class GlobalPointConditionFactory implements PlayerConditionFactory, PlayerlessConditionFactory {
-
-    /**
-     * The global data.
-     */
-    private final GlobalData globalData;
+public class GlobalPointConditionFactory extends DefaultPointConditionFactory implements PlayerConditionFactory, PlayerlessConditionFactory {
 
     /**
      * Creates the global point factory.
@@ -27,23 +20,16 @@ public class GlobalPointConditionFactory implements PlayerConditionFactory, Play
      * @param globalData the global data
      */
     public GlobalPointConditionFactory(final GlobalData globalData) {
-        this.globalData = globalData;
+        super(profile -> globalData.points());
     }
 
     @Override
     public PlayerCondition parsePlayer(final Instruction instruction) throws QuestException {
-        return new NullableConditionAdapter(parse(instruction));
+        return new NullableConditionAdapter(parseInstruction(instruction));
     }
 
     @Override
     public PlayerlessCondition parsePlayerless(final Instruction instruction) throws QuestException {
-        return new NullableConditionAdapter(parse(instruction));
-    }
-
-    private GlobalPointCondition parse(final Instruction instruction) throws QuestException {
-        final Argument<String> category = instruction.packageIdentifier().get();
-        final Argument<Number> count = instruction.number().get();
-        final FlagArgument<Boolean> equal = instruction.bool().getFlag("equal", true);
-        return new GlobalPointCondition(globalData, category, count, equal);
+        return new NullableConditionAdapter(parseInstruction(instruction));
     }
 }